Revert BungeeColor-Regular.ttf to COLRv0 only
The git history of this file was a bit messy so I couldn't simply git revert, but replaced the current BungeeColor-Regular.ttf with the upstream one found on djrrb/Bungee repository.

This effectively restores the font to its COLRv0 state, no additional color tables added by the maximum_color tool.