Re-run maximum_color on BungeeSpice-Regular.ttf to fix issue with colored .notdef glyph being reshuffled (#5171)
There was an issue in the maximum_color tool that affected BungeeSpice-Regular.ttf, see https://github.com/googlefonts/nanoemoji/issues/427

Here I took the original BungeeSpice from upstream djrrb repository and run it through the latest maximum_color (v0.14.3).

The font has the (colored) .notdef glyph at index 0, and space character is again mapped to 0x20 in cmap as it should be. The space was currently un-cmapped (for it had been moved to index 0) and thus as-if missing.