italic.textproto reduce precision from -1 to 0
Since we don't support the ital axis, in fact, and require separate roman and italic files at the moment (due to concerns eg in https://arrowtype.github.io/vf-slnt-test), then we should reflect that actual implementation in the axis definition. 

Since we have zero families with an ital axis, this should be a no-op when pushed; but this is a very peculiar move, and may have unexpected consequences when pushed.