Addition of exemplar character sets for Latin-based African Languages
Additional 313 exemplar character sets have been added to support Latin-based African languages. This is commit is a precursor to performing shaperglot analysis of the GF library and Google Sans.

Data is sourced from ScriptSource which has been reorganized and modified to include uppercase letters and isolate marks. Of the 1913 African language tags with no exemplar char data in gflang the these are the only ones with data on ScriptSource.