Update method of installing FontBakery
FontBakery now requires the name of the desired profile to be passed as an extra when installing it.
For google/fonts, we run the `Google Fonts` profile, so the installation
is done with `pip install --pre fontbakery[googlefonts]`

Without passing extras, only the dependencies needed by the Universal and
OpenType profiles (and any vendor-specific profile not requiring
additional deps) are installed.