Update families.csv
Two fonts, Carter One and Festive, had duplicate entries for /Seasonal/Hanukkah and those were accidentally converted to one for Hanukkah and one for Kwanzaa.  This PR removes the Kwanzaa ones.

Also, Zen Tokyo Zoo had a tag added for "Kwanzaa" and it was changed to "/Seasonal" and it needs to be "/Seasonal/Kwanzaa".

Using Marc's tool to make these changes caused the file to be sorted again.  Some of the "/Seasonal/Kwanzaa" were not sorted properly, but will be after this PR is merged.