Update to_production.txt (#9128)
Removed the entries that were pushed to prod.  At this point, they have been pushed to our Canary instances and should be fully pushed to prod on Mon Mar 3.

Four fonts:
   Buenard
   Cormorant Garamond
   Cormorant Infant
   Merriweather
are being converted from statics to a VF and this requires some extra work on the Google side as some other parts of Google reference the static fonts.